+++ tests/testthat/test-aet01.R
@@ -15,6 +15,7 @@ test_that("aet01 can handle all NA values", {
 })
 
 test_that("aet01 can handle some NA values", {
+  skip_on_os("windows")
   new_aesdth <- c(NA, "", as.character(syn_data$adae$AESDTH[-c(1, 2)]))
   new_aeser <- c(NA, "", as.character(syn_data$adae$AESER[-c(1, 2)]))
 
@@ -30,12 +31,14 @@ test_that("aet01 can handle some NA values", {
 })
 
 test_that("aet01 can use custom anl_vars", {
+  skip_on_os("windows")
   proc_data <- syn_data
   res <- expect_silent(run(aet01, proc_data, anl_vars = list(safety_var = "FATAL")))
   expect_snapshot(cat(export_as_txt(res, lpp = 100)))
 })
 
 test_that("aet01 fails on incomplete data input", {
+  skip_on_os("windows")
   proc_data <- syn_data
   proc_data$adae <- proc_data$adae %>%
     mutate(AESER = NULL)
@@ -48,6 +51,7 @@ test_that("aet01 fails on incomplete data input", {
 })
 
 test_that("aet01 can use custom medconcept_var", {
+  skip_on_os("windows")
   proc_data <- syn_data
   proc_data$adae$SMQ01 <- with_label(proc_data$adae$SMQ01NAM != "", "SMQ 01")
   res <- expect_silent(
@@ -66,6 +70,7 @@ test_that("aet01 can use custom medconcept_var", {
 })
 
 test_that("aet01 can conditionally show the number of withdrawal", {
+  skip_on_os("windows")
   expect_silent(res <- run(aet01, syn_data, arm_var = "ARM", show_wd = FALSE))
   expect_snapshot(cat(export_as_txt(res, lpp = 100)))
 })