+++ tests/testthat/test-dst01.R
@@ -12,6 +12,7 @@ test_that("dst01 works as expected when all data are NA", {
 })
 
 test_that("dst01 can handle all NA values in DCSREAS", {
+  skip_on_os("windows")
   proc_data <- syn_data
   proc_data$adsl <- proc_data$adsl %>%
     mutate(
@@ -23,6 +24,7 @@ test_that("dst01 can handle all NA values in DCSREAS", {
 })
 
 test_that("dst01 can handle some NA values", {
+  skip_on_os("windows")
   proc_data <- syn_data
   proc_data$adsl[1:2, c("EOSSTT", "DCSREAS")] <- NA
   res <- expect_silent(run(dst01, proc_data))
@@ -30,6 +32,7 @@ test_that("dst01 can handle some NA values", {
 })
 
 test_that("dst01 can handle missing levels in status_var", {
+  skip_on_os("windows")
   proc_data <- syn_data
   proc_data$adsl <- proc_data$adsl %>%
     mutate(
@@ -40,16 +43,19 @@ test_that("dst01 can handle missing levels in status_var", {
 })
 
 test_that("dst01 can create variants", {
+  skip_on_os("windows")
   res <- expect_silent(run(dst01, syn_data, detail_vars = list(Ongoing = "STDONS", Discontinued = "DCSREAS")))
   expect_snapshot(cat(export_as_txt(res, lpp = 100)))
 })
 
 test_that("dst01 can create variants", {
+  skip_on_os("windows")
   res <- expect_silent(run(dst01, syn_data, detail_vars = list(Discontinued = c("DCSREASGP", "DCSREAS"))))
   expect_snapshot(cat(export_as_txt(res, lpp = 100)))
 })
 
 test_that("dst01 can create variants", {
+  skip_on_os("windows")
   res <- expect_silent(
     run(dst01, syn_data, detail_vars = list(Discontinued = c("DCSREASGP", "DCSREAS")), trt_status_var = "EOTSTT")
   )