+++ tests/testthat/test-egt02.R
@@ -8,6 +8,7 @@ test_that("egt02_1 functions with default argument value return expected result 
 })
 
 test_that("egt02_2 functions with default value return expected result with test data", {
+  skip_on_os("windows")
   pre_data <- egt02_pre(syn_data)
   raw_res <- egt02_2_main(pre_data)
   res <- egt02_post(raw_res)